:recycle: : 문자인증 리팩토링
