+++ Match-Api/src/main/java/com/example/matchapi/user/utils/AuthHelper.java
@@ -54,37 +54,6 @@ public class AuthHelper {
         }
     }
 
-    public String createRandomNumber() {
-        Random rand  = new Random();
-        String numStr = "";
-        for(int i=0; i<6; i++) {
-            String ran = Integer.toString(rand.nextInt(10));
-            numStr+=ran;
-        }
-
-        return numStr;
-    }
-
-    /*
-    public void sendSms(String phone, String number) throws CoolsmsException {
-        String api_key = coolSmsProperties.getApi(); // 발급받은 api_key
-        String api_secret = coolSmsProperties.getSecret(); // 발급받은 api_secret
-        Message coolsms = new Message(api_key, api_secret);
-
-
-
-        HashMap<String, String> params = new HashMap<String, String>();
-        params.put("to",phone);    // 수신전화번호 (ajax로 view 화면에서 받아온 값으로 넘김)
-        params.put("from", coolSmsProperties.getSender());    // 발신전화번호. 테스트시에는 발신,수신 둘다 본인 번호로 하면 됨
-        params.put("type", "sms");
-        params.put("text", "인증번호는 [" + number + "] 입니다.");
-
-        coolsms.send(params); // 메시지 전송
-
-    }
-
-     */
-
     @FunctionalInterface
     private interface GenderResolver {
         Gender resolveGender(String gender);
